+++ coldfront/core/project/utils_/renewal_survey_backend/backends/google.py
@@ -1,7 +1,7 @@
 from coldfront.core.project.utils_.renewal_survey_backend.backends.base import BaseRenewalSurveyBackend
-from coldfront.core.project.utils_.renewal_utils import get_gspread_wks
-from coldfront.core.project.utils_.renewal_utils import get_renewal_survey
-from coldfront.core.project.utils_.renewal_utils import gsheet_column_to_index
+from coldfront.core.project.utils_.renewal_survey_utils import get_gspread_wks
+from coldfront.core.project.utils_.renewal_survey_utils import get_renewal_survey
+from coldfront.core.project.utils_.renewal_survey_utils import gsheet_column_to_index
 
 from django.core.exceptions import ValidationError
 
@@ -100,35 +100,3 @@ class GoogleRenewalSurveyBackend(BaseRenewalSurveyBackend):
                 PARAMETER_BASE_TWO + value
         return url
     
-
-def make_form_url(self, survey_data, dictionary):
-        """ This function creates the Google Form URL with appropriate
-         pre-filled parameters. """
-        BASE_URL_ONE = 'https://docs.google.com/forms/d/e/'
-        BASE_URL_TWO = '/viewform?usp=pp_url'
-        url = BASE_URL_ONE + survey_data['form_id'] + BASE_URL_TWO
-
-        PARAMETER_BASE_ONE = '&entry.'
-        PARAMETER_BASE_TWO = '='
-
-        question_ids_dict = survey_data['form_question_ids']
-        for parameter in question_ids_dict.keys():
-            value = ''
-            if parameter == 'allocation_period':
-                value = dictionary['allocation_period'].name
-            elif parameter == 'pi_name':
-                value = dictionary['PI'].user.first_name + '+' + \
-                    dictionary['PI'].user.last_name
-            elif parameter == 'pi_username':
-                value = dictionary['PI'].user.username
-            elif parameter == 'project_name':
-                value = self.project_obj.name
-            elif parameter == 'requester_name':
-                value = self.request.user.first_name + '+' + \
-                    self.request.user.last_name
-            elif parameter == 'requester_username':
-                value = self.request.user.username
-            value = value.replace(' ', '+')
-            url += PARAMETER_BASE_ONE + question_ids_dict[parameter] + \
-                PARAMETER_BASE_TWO + value
-        return url
