+++ coldfront/core/project/views_/renewal_views/request_views.py
@@ -26,7 +26,7 @@ from coldfront.core.project.utils_.renewal_utils import has_non_denied_renewal_r
 from coldfront.core.project.utils_.renewal_utils import send_new_allocation_renewal_request_admin_notification_email
 from coldfront.core.project.utils_.renewal_utils import send_new_allocation_renewal_request_pi_notification_email
 from coldfront.core.project.utils_.renewal_utils import send_new_allocation_renewal_request_pooling_notification_email
-from coldfront.core.project.utils_.renewal_utils import get_renewal_survey
+from coldfront.core.project.utils_.renewal_survey_utils import get_renewal_survey
 from coldfront.core.project.utils_.renewal_survey_backend import get_survey_url
 from coldfront.core.resource.models import Resource
 from coldfront.core.resource.utils import get_primary_compute_resource
@@ -598,6 +598,7 @@ class AllocationRenewalRequestView(LoginRequiredMixin, UserPassesTestMixin,
     def make_form_url(self, survey_data, dictionary):
         """Get pre-filled survey link for user."""
         dictionary['requester'] = self.request.user
+        dictionary['project_name'] = dictionary['requested_project'].name
         return get_survey_url(survey_data, dictionary)
 
 class AllocationRenewalRequestUnderProjectView(LoginRequiredMixin,
